Fix bioav differentiation when rate!=0 ss!=0 and no other DCPs
As the title says, this is a very very specific case. Essentially what was happening was that if rate!=0 and ss!=0, you could still have a coupling in the steady state off timings due to bioav, since the amount of the dose would change where the endpoints of the SS doses would lie, so you need the time coupling. But the time derivative coupling after https://github.com/PumasAI/Pumas.jl/pull/912 only existed when event times were dual (i.e. to fix lags, rate, and duration), so the simplest idea is to make time dual when bioav is dual since there is a good chance that it will have time-dependent properties. This just imparts the dual-ness by multiplying by bioav/bioav