Add continuous-time Markov models
This adds the ability to utilize continuous-time Markov models. Most of the weird issues come from the fact that in this case, Pumas needs to modify the internal `prob.prob` other than the `prob` directly since it's a `JumpProblem`, and the jump pieces need to be reset at dosing. However, since the model is memory-less, simply a refresh of the rates + a new exponential random variable is sufficient to retain the correct statistics, so a hook was created for Pumas to do the proper resetting.

This superscedes https://github.com/PumasAI/Pumas.jl/pull/472 and fixes https://github.com/PumasAI/Pumas.jl/issues/136
remove shows