Fix analytical repeated events and test both diffeq and analytic repeats
Fixes https://github.com/PumasAI/Pumas.jl/issues/882 and explicitly adds a test for both analytical and numerical based repeated events. The solution is actually quite simple: just always increment the dose, and explicitly zero it when you move to a new time. This doesn't intersect with the steady state handling because SS's must be unique and so nothing needs to change there (and that explicitly has a test for steady states hitting ii).