mtime mutability: solver-time mutable parameters
Introduces `@cache` for declaring mutable components of the generated pre function which can be changed using the callback system. This allows for callbacks to change parts of the dynamics, giving a flexible mtime system. This is done by creating a type and doing a call-overloading, and using that call-overloaded type as the pre function.

Fixes https://github.com/PumasAI/Pumas.jl/issues/907 .

In theory this could introduce a regression since the macro is now defining a type, which means it would need to be used at the top level. We can make this function-safe by doing some trickery like evaling into module and then grabbing that definition from the module, but I hope it doesn't need to come to that...