Handle evid == 3 and evid == 4 through data parsing (#1101)
* Handle evid == 3 and evid == 4 through data parsing

When parsing data, go through the data sequentially. If you hit evid == 3 or evid == 4, revert the TAD form by making an offset at that time, and shift all further events by that amount.

Note that supporting evid == 3 and evid == 4 requires that we get rid of checks for monotonicity in the data, because obviously TAD-based data are not going to be monotonic. It also means that (time,id) pairs are no longer unique, as demonstrated by the test. So these erroneous checks were deleted.

The test case then checks against an analytical solution for what the values should be to make sure that the events are realigned, and then uses the analytical solution as the observation values in the deviance to make sure the observations are realigned. If everything is perfect then the deviance should be infinity, so with numerical tolerances set it just checks if it's absurdly high.

* allow nothing time

* remove chronological test

* add some monotonicity checks back

* index the dataframe properly

* add tad parsing off switch

* fix time monotonicity check

* Sort datasets where events have been appended

* Use TAD in repeated time-to-event

* Update src/data_parsing/io.jl

Co-authored-by: Andreas Noack <andreas@noack.dk>

* don't error on first event reset

* Update test/core/parsing_tests.jl

* Fix repeated time-to-event test

Co-authored-by: Andreas Noack <andreas@noack.dk>