Add parallel calls for multiple parameters in simobs (#1033)
* Add parallel calls for multiple parameters in simobs

Fixes #1000 

Makes it so that if you pass N parameters for 1 subject, you get N solutions, 1 for each of the parameters, and parallelizes it. So N parameters and N subjects is symmetric, and errors check to make sure they are the same size if you pass in both. Using FillArrays allows us to do this without copying, so we should probably use this more.

* expose Population,Array of Parameters interface

* fix param and randeffs generation for ensembling

* make sure to pass on randeffs in the parallelized versions

* Adjust negativebinomial.jl tests for new simobs

* Adjust tvcov.jl tests to new simobs

* Adjust mdv.jl tests to new simobs

* Adjust dosecontrolparameters.jl test to new simobs

* Adjust hcv.jl test to new simobs

Co-authored-by: Andreas Noack <andreas@noack.dk>